// src/data/models/domain/filtertypes.ts
export enum filterTypes{
  QuoteId = "Quote ID",
  QuotedAfterDate = "After",
  QuotedBeforeDate = "Before",
  MinimumPrice = "Min Price",
  MaximumPrice = "Max Price",
  FirstName = "First Name",
  LastName = "Last Name",
  Phone = "Phone No",
  Email = "Email",
  City = "City",
  State = "State",
  Zip = "Zip"
}